Restored the removed assignment operators for timers from integers and PTimeInterval, all sorts of things stopped working!
Also added copy constructor and assignment operator for PTimer, as the operating system default causes all sorts of problems. New implementation is equivalent of setting the new timer to the right hand sides "time remaining" value.

Removed a mutex usage that was causing deadlock, code it used to protect is only ever touched within one thread, so not needed.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@22741 023b2edf-31b2-4de3-b41e-bca80c47788f