Added asynchronous notifier. This is similar to a PNotifier but the execution of the call back is queued and delayed until an execute function is called. This allows all call backs to be executed in a single thread (or thread pool) no matter what thread initiated the notification. This is especially useful in GUI environment where touching the windowing system is fraught with deadlocks unless always done by the "main" process thread.
While in the process did a massive tidying of the existing "smart" notifier, which was renamed to a more meaningful "validated" notifier. This is where the target object is validated before a synchronous call is made to it, thus avoiding crashes when the lifetime of the target is not guaranteed to be longer than the notifier.

Finally, templated then notifier list class so is type safe like all the other notifiers.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@26296 023b2edf-31b2-4de3-b41e-bca80c47788f