Hack to fix for latest GLIBC static object initialisation problem. It appears that when libpt.so does a dlopen() of a library that uses libpt.so, say v4l2_pwplugin.so for example, then all of the static object constructors are run AGAIN.
This may or may not be a problem depending on the static, for us, it is only a problem if a singleton has been instantiated before the dlopen(). If not all is well, a NULL poointer is set back to NULL. It is has then confusion reigns and crashes can ensue.

Not having any idea how to prevent this behaviour short of completely rewriting the audo/video device plugin system (tempting, but no time ...) this is a complete hack so that the plugins are loaded as early as possible to be (so far) before any singletons are instantiated.

This has the potential of going wrong in the future but, as I said, no idea how to fix it outside of glibc itself!


git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/branches/v2_10@26627 023b2edf-31b2-4de3-b41e-bca80c47788f